+++ Source/LuaBridge/Map.h
@@ -18,7 +18,7 @@ struct Stack <std::map <K, V> >
 
   static void push(lua_State* L, const Map& map)
   {
-    lua_createtable (L, 0, map.size ());
+    lua_createtable (L, 0, static_cast <int> (map.size ()));
     typedef typename Map::const_iterator ConstIter;
     for (ConstIter i = map.begin(); i != map.end(); ++i)
     {
